+++ sword-tests/src/request/cookies.rs
@@ -16,7 +16,7 @@ impl OnRequest for SetCookieMw {
 
         cookies.add(cookie);
 
-        next!(req, next)
+        req.run(next).await
     }
 }