+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Sword web framework changelog
 
-## Next Release
+## [Unreleased]
+
+### Added
+
+- New `run` method on `Request` struct. See `Changed` section for more details.
+
+### Changed
+
+- The `next!` macro has been deprecated. Instead, use the `req.run(next).await` method to pass control to the next middleware or handler in the chain. This change removes the need for a macro to do "magic" and makes the code more explicit and easier to understand.
+
+
+## [0.2.0]
 
 ### Added